fix: address loop-panel PR review (BPM, save_loop, hydration)
Validate finite BPM > 0 for loop ops, preserve loop label/color on update,
and move track hydration off build side effects.

+++ apps/gui-flutter/lib/mixer/deck_loop_panel.dart
@@ -20,8 +20,7 @@ int stepAutoLoopBeats(int beats, int delta) {
 }
 
 /// Slot used for save/recall of the loop length chip (Tauri: beat index, max 7).
-int autoLoopSlotForBeats(int beats) =>
-    autoLoopBeatIndex(beats).clamp(0, 7).toInt();
+int autoLoopSlotForBeats(int beats) => autoLoopBeatIndex(beats).clamp(0, 7);
 
 /// Saved loop under [positionMs]: playhead inside `[in, out]`, tightest span,
 /// then lowest slot.
